Fix module layout; remove student's directory
