+++ go.mod
@@ -1,0 +1,3 @@
+module urlshort
+
+go 1.21.3
