--- students/dennisvis/paths.json
@@ -1,10 +1,0 @@
-[
-  {
-    "path": "/urlshort",
-    "url": "https://github.com/gophercises/urlshort"
-  },
-  {
-    "path": "/urlshort-final",
-    "url": "https://github.com/gophercises/urlshort/tree/solution"
-  }
-]
