--- students/dennisvis/paths.yml
@@ -1,4 +1,0 @@
-- path: /urlshort
-  url: https://github.com/gophercises/urlshort
-- path: /urlshort-final
-  url: https://github.com/gophercises/urlshort/tree/solution
