--- students/dimdiden/map.json
@@ -1,10 +1,0 @@
-[
-  {
-    "path": "/mobile",
-    "url" : "https://mobile-review.com/"
-  },
-  {
-    "path": "/pikabu",
-    "url" : "https://pikabu.ru/"
-  }
-]
