--- students/emrekp/handler_impls.go
@@ -1,60 +1,0 @@
-package urlshort
-
-import (
-	"gopkg.in/yaml.v2"
-	"net/http"
-)
-
-// MapHandler will return an http.HandlerFunc (which also
-// implements http.Handler) that will attempt to map any
-// paths (keys in the map) to their corresponding URL (values
-// that each key in the map points to, in string format).
-// If the path is not provided in the map, then the fallback
-// http.Handler will be called instead.
-func MapHandler(paths map[string]string, fallback http.Handler) http.HandlerFunc {
-	return func(writer http.ResponseWriter, request *http.Request) {
-		url := paths[request.URL.Path]
-		if url != "" {
-			http.Redirect(writer, request, url, http.StatusFound)
-			return
-		}
-
-		fallback.ServeHTTP(writer, request)
-	}
-}
-
-type YAMLMap struct {
-	Path string `yaml:"path"`
-	Url  string `yaml:"url"`
-}
-
-// YAMLHandler will parse the provided YAML and then return
-// an http.HandlerFunc (which also implements http.Handler)
-// that will attempt to map any paths to their corresponding
-// URL. If the path is not provided in the YAML, then the
-// fallback http.Handler will be called instead.
-//
-// YAML is expected to be in the format:
-//
-//     - path: /some-path
-//       url: https://www.some-url.com/demo
-//
-// The only errors that can be returned all related to having
-// invalid YAML data.
-//
-// See MapHandler to create a similar http.HandlerFunc via
-// a mapping of paths to urls.
-func YAMLHandler(yml []byte, fallback http.Handler) (http.HandlerFunc, error) {
-	var ymlMap []YAMLMap
-	err := yaml.Unmarshal(yml, &ymlMap)
-	if err != nil {
-		return nil, err
-	}
-
-	paths := make(map[string]string)
-	for _, y := range ymlMap {
-		paths[y.Path] = y.Url
-	}
-
-	return MapHandler(paths, fallback), nil
-}
