--- students/hackeryarn/urls.json
@@ -1,10 +1,0 @@
-[
-    {
-        "path": "/urlShort",
-        "url": "https://github.com/gophercises/urlshort"
-    },
-    {
-        "path": "/urlshortFinal",
-        "url": "https://github.com/gophercises/urlshort/tree/solution"
-    }
-]
