--- students/liikt/storage/map.json
@@ -1,4 +1,0 @@
-[
-    {"path": "/yt", "url":"https://www.youtube.com/"},
-    {"path": "/tr", "url":"http://play.typeracer.com/"}
-]
